requirements.txt
discord.py
yt-dlp
sqlite3
